Began the long and painstaking process of rewriting.
