+++ legacy/scripts/convert_logo.py
