+++ legacy/setup.sh
