+++ mypy.ini
@@ -1,0 +1,3 @@
+[mypy]
+ignore_missing_imports = True
+files = airgui, airscript, airweb, airscript-ng.py
