+++ tox.ini
@@ -1,0 +1,3 @@
+[flake8]
+max-line-length=120
+exclude=__pycache__
